+++ assertions/src/main/java/io/domainlifecycles/assertion/DomainAssertions.java
@@ -51,7 +51,7 @@ import java.util.Optional;
  * <p>
  * If an assertion fails a DomainAssertionException is thrown.
  *
- * @author Mario Herb
+ * @author Mario Herb, Philipp Holz
  */
 public class DomainAssertions {
 
@@ -64,10 +64,21 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void equals(Object anObject1, Object anObject2, String aMessage) {
-        if ((anObject1 != null && !anObject1.equals(anObject2))
-            || (anObject1 == null && anObject2 != null)) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.equals(anObject1, anObject2, aMessage);
+    }
+
+    /**
+     * Compares two objects for equality. Two null object references are considered to be equal.
+     *
+     * @param anObject1 checked
+     * @param anObject2 compare value
+     * @param runtimeException exception that will be thrown
+     * @throws RuntimeException if check failed
+     */
+    public static void equals(Object anObject1, Object anObject2, RuntimeException runtimeException) {
+
+        DomainAssertionsEquals.equals(anObject1, anObject2, runtimeException);
     }
 
     /**
@@ -80,15 +91,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void optionalEquals(Optional anObjectOptional, Object anObject2, String aMessage) {
-        if (anObjectOptional == null) {
-            throw new IllegalArgumentException("The first parameter must not be null!");
-        }
-        if (anObjectOptional.isPresent()) {
-            Object anObject1 = anObjectOptional.get();
-            equals(anObject1, anObject2, aMessage);
-        } else if (anObjectOptional.isEmpty() && anObject2 != null) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.optionalEquals(anObjectOptional, anObject2, aMessage);
     }
 
     /**
@@ -100,9 +104,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void equals(int aNumber1, int aNumber2, String aMessage) {
-        if (aNumber1 != aNumber2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.equals(aNumber1, aNumber2, aMessage);
     }
 
     /**
@@ -114,9 +117,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void equals(long aNumber1, long aNumber2, String aMessage) {
-        if (aNumber1 != aNumber2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.equals(aNumber1, aNumber2, aMessage);
     }
 
     /**
@@ -128,9 +130,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void equals(byte aNumber1, byte aNumber2, String aMessage) {
-        if (aNumber1 != aNumber2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.equals(aNumber1, aNumber2, aMessage);
     }
 
     /**
@@ -142,9 +143,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void equals(short aNumber1, short aNumber2, String aMessage) {
-        if (aNumber1 != aNumber2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.equals(aNumber1, aNumber2, aMessage);
     }
 
     /**
@@ -156,9 +156,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void equals(double aNumber1, double aNumber2, String aMessage) {
-        if (aNumber1 != aNumber2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.equals(aNumber1, aNumber2, aMessage);
     }
 
     /**
@@ -170,9 +169,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void equals(float aNumber1, float aNumber2, String aMessage) {
-        if (aNumber1 != aNumber2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.equals(aNumber1, aNumber2, aMessage);
     }
 
     /**
@@ -184,9 +182,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void equals(char aChar1, char aChar2, String aMessage) {
-        if (aChar1 != aChar2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.equals(aChar1, aChar2, aMessage);
     }
 
     /**
@@ -198,9 +195,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void equals(boolean aBool1, boolean aBool2, String aMessage) {
-        if (aBool1 != aBool2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.equals(aBool1, aBool2, aMessage);
     }
 
     /**
@@ -212,10 +208,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void notEquals(Object anObject1, Object anObject2, String aMessage) {
-        if ((anObject1 != null && anObject1.equals(anObject2))
-            || (anObject1 == null && anObject2 == null)) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.notEquals(anObject1, anObject2, aMessage);
     }
 
     /**
@@ -228,15 +222,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void optionalNotEquals(Optional anOptionalObject1, Object anObject2, String aMessage) {
-        if (anOptionalObject1 == null) {
-            throw new IllegalArgumentException("The first parameter must not be null!");
-        }
-        if (anOptionalObject1.isPresent()) {
-            Object anObject1 = anOptionalObject1.get();
-            notEquals(anObject1, anObject2, aMessage);
-        } else if (anOptionalObject1.isEmpty() && anObject2 == null) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.optionalNotEquals(anOptionalObject1, anObject2, aMessage);
     }
 
     /**
@@ -248,9 +235,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void notEquals(int aNumber1, int aNumber2, String aMessage) {
-        if (aNumber1 == aNumber2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.notEquals(aNumber1, aNumber2, aMessage);
     }
 
     /**
@@ -263,9 +249,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void notEquals(long aNumber1, long aNumber2, String aMessage) {
-        if (aNumber1 == aNumber2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.notEquals(aNumber1, aNumber2, aMessage);
     }
 
     /**
@@ -277,9 +262,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void notEquals(byte aNumber1, byte aNumber2, String aMessage) {
-        if (aNumber1 == aNumber2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.notEquals(aNumber1, aNumber2, aMessage);
     }
 
     /**
@@ -291,9 +275,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void notEquals(short aNumber1, short aNumber2, String aMessage) {
-        if (aNumber1 == aNumber2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.notEquals(aNumber1, aNumber2, aMessage);
     }
 
     /**
@@ -305,9 +288,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void notEquals(double aNumber1, double aNumber2, String aMessage) {
-        if (aNumber1 == aNumber2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.notEquals(aNumber1, aNumber2, aMessage);
     }
 
     /**
@@ -319,9 +301,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void notEquals(float aNumber1, float aNumber2, String aMessage) {
-        if (aNumber1 == aNumber2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.notEquals(aNumber1, aNumber2, aMessage);
     }
 
     /**
@@ -333,9 +314,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void notEquals(char aChar1, char aChar2, String aMessage) {
-        if (aChar1 == aChar2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.notEquals(aChar1, aChar2, aMessage);
     }
 
     /**
@@ -347,9 +327,8 @@ public class DomainAssertions {
      * @throws DomainAssertionException if check failed
      */
     public static void notEquals(boolean aBool1, boolean aBool2, String aMessage) {
-        if (aBool1 == aBool2) {
-            throw new DomainAssertionException(aMessage);
-        }
+
+        DomainAssertionsEquals.notEquals(aBool1, aBool2, aMessage);
     }
 
     /**