+++ assertions/src/test/java/io/domainlifecycles/assertion/CustomAssertionException.java
@@ -1,0 +1,18 @@
+package io.domainlifecycles.assertion;
+
+class CustomAssertionException extends RuntimeException {
+
+    private final CustomAssertionExceptionCode errorCode;
+
+    CustomAssertionException(final String message, final CustomAssertionExceptionCode errorCode) {
+
+        super(message);
+        this.errorCode = errorCode;
+    }
+
+    CustomAssertionExceptionCode exceptionCode() {
+
+        return errorCode;
+    }
+}
+