+++ assertions/src/test/java/io/domainlifecycles/assertion/CustomAssertionExceptionCode.java
@@ -1,0 +1,6 @@
+package io.domainlifecycles.assertion;
+
+enum CustomAssertionExceptionCode {
+
+    CUSTOM_ASSERTION_EXCEPTION_CODE_1,
+}