+++ assertions/src/test/java/io/domainlifecycles/assertion/CustomAssertionExceptions.java
@@ -1,0 +1,9 @@
+package io.domainlifecycles.assertion;
+
+class CustomAssertionExceptions {
+
+    static CustomAssertionException defaultCustomAssertionException() {
+
+        return new CustomAssertionException("Failed", CustomAssertionExceptionCode.CUSTOM_ASSERTION_EXCEPTION_CODE_1);
+    }
+}