+++ assertions/src/test/java/io/domainlifecycles/assertion/TestAssertionsEquals.java
@@ -7,6 +7,9 @@ import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.assertj.core.api.AssertionsForClassTypes.catchThrowableOfType;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class TestAssertionsEquals {
 
@@ -14,119 +17,146 @@ public class TestAssertionsEquals {
     class TestEquals {
 
         @Test
-        public void testObjectEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.equals("A", "B", "Failed")).isInstanceOf(
+        void testObjectEqualsFailCustomException() {
+
+            // org.junit.jupiter.api.assertThrows -> preferred lib?
+            final var exception = assertThrows(
+                CustomAssertionException.class,
+                () -> DomainAssertionsEquals.equals(
+                    "A",
+                    "B",
+                    CustomAssertionExceptions.defaultCustomAssertionException()
+                ));
+
+            assertThat(exception.exceptionCode())
+                .isEqualTo(CustomAssertionExceptionCode.CUSTOM_ASSERTION_EXCEPTION_CODE_1);
+
+            // this method is changed in a new assertj version, reversing the parameters.
+            // we should decide, which api we want to use. org.junit.jupiter or org.assertj
+            // org.assertj.core.api.catchThrowableOfType
+            final var exception2 = catchThrowableOfType(
+                () -> DomainAssertionsEquals.equals("A", "B", CustomAssertionExceptions.defaultCustomAssertionException()),
+                CustomAssertionException.class
+            );
+            assertThat(exception2.exceptionCode())
+                .isEqualTo(CustomAssertionExceptionCode.CUSTOM_ASSERTION_EXCEPTION_CODE_1);
+        }
+
+        @Test
+        void testObjectEqualsFail() {
+            // org.assertj.core.api.assertThatThrownBy -> replace with org.junit.jupiter.api.assertThrows
+            assertThatThrownBy(() -> DomainAssertionsEquals.equals("A", "B", "Failed"))
+                .isInstanceOf(DomainAssertionException.class);
+        }
+
+        @Test
+        void testObjectEqualsNullFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.equals(null, "B", "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testObjectEqualsNullFail() {
-            assertThatThrownBy(() -> DomainAssertions.equals(null, "B", "Failed")).isInstanceOf(
+        void testObjectEqualsNullCompareFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.equals("A", null, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testObjectEqualsNullCompareFail() {
-            assertThatThrownBy(() -> DomainAssertions.equals("A", null, "Failed")).isInstanceOf(
+        void testObjectEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.equals("A", "A", "Failed"));
+        }
+
+        @Test
+        void testObjectEqualsBothNullOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.equals(null, null, "Failed"));
+        }
+
+        @Test
+        void testIntEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.equals(5, 6, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testObjectEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.equals("A", "A", "Failed"));
+        void testIntEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.equals(3, 3, "Failed"));
         }
 
         @Test
-        public void testObjectEqualsBothNullOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.equals(null, null, "Failed"));
-        }
-
-        @Test
-        public void testIntEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.equals(5, 6, "Failed")).isInstanceOf(
+        void testLongEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.equals(5l, 6l, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testIntEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.equals(3, 3, "Failed"));
+        void testLongEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.equals(3l, 3l, "Failed"));
         }
 
         @Test
-        public void testLongEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.equals(5l, 6l, "Failed")).isInstanceOf(
+        void testByteEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.equals((byte) 5, (byte) 6, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testLongEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.equals(3l, 3l, "Failed"));
+        void testByteEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.equals((byte) 3, (byte) 3, "Failed"));
         }
 
         @Test
-        public void testByteEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.equals((byte) 5, (byte) 6, "Failed")).isInstanceOf(
+        void testShortEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.equals((short) 5, (short) 6, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testByteEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.equals((byte) 3, (byte) 3, "Failed"));
+        void testShortEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.equals((short) 3, (short) 3, "Failed"));
         }
 
         @Test
-        public void testShortEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.equals((short) 5, (short) 6, "Failed")).isInstanceOf(
+        void testDoubleEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.equals(5.0, 6.0, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testShortEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.equals((short) 3, (short) 3, "Failed"));
+        void testDoubleEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.equals(3.0, 3.0, "Failed"));
         }
 
         @Test
-        public void testDoubleEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.equals(5.0, 6.0, "Failed")).isInstanceOf(
+        void testFloatEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.equals((float) 5.0, (float) 6.0, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testDoubleEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.equals(3.0, 3.0, "Failed"));
+        void testFloatEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.equals((float) 3.0, (float) 3.0, "Failed"));
         }
 
         @Test
-        public void testFloatEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.equals((float) 5.0, (float) 6.0, "Failed")).isInstanceOf(
+        void testCharEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.equals('a', 'b', "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testFloatEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.equals((float) 3.0, (float) 3.0, "Failed"));
+        void testCharEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.equals('x', 'x', "Failed"));
         }
 
         @Test
-        public void testCharEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.equals('a', 'b', "Failed")).isInstanceOf(
+        void testBooleanEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.equals(true, false, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testCharEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.equals('x', 'x', "Failed"));
-        }
-
-        @Test
-        public void testBooleanEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.equals(true, false, "Failed")).isInstanceOf(
-                DomainAssertionException.class);
-        }
-
-        @Test
-        public void testBooleanEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.equals(false, false, "Failed"));
+        void testBooleanEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.equals(false, false, "Failed"));
         }
     }
 
@@ -134,118 +164,118 @@ public class TestAssertionsEquals {
     class TestNotEquals {
 
         @Test
-        public void testObjectNotEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.notEquals("A", "A", "Failed")).isInstanceOf(
+        void testObjectNotEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.notEquals("A", "A", "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testObjectNotEqualsBothNullFail() {
-            assertThatThrownBy(() -> DomainAssertions.notEquals(null, null, "Failed")).isInstanceOf(
+        void testObjectNotEqualsBothNullFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.notEquals(null, null, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testObjectNotEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.notEquals("A", "A1", "Failed"));
+        void testObjectNotEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.notEquals("A", "A1", "Failed"));
         }
 
         @Test
-        public void testObjectNotEqualsNullOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.notEquals(null, "A1", "Failed"));
+        void testObjectNotEqualsNullOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.notEquals(null, "A1", "Failed"));
         }
 
         @Test
-        public void testObjectNotEqualsNullParameterOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.notEquals("A", null, "Failed"));
+        void testObjectNotEqualsNullParameterOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.notEquals("A", null, "Failed"));
         }
 
         @Test
-        public void testIntNotEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.notEquals(5, 5, "Failed")).isInstanceOf(
+        void testIntNotEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.notEquals(5, 5, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testIntNotEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.notEquals(4, 3, "Failed"));
+        void testIntNotEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.notEquals(4, 3, "Failed"));
         }
 
         @Test
-        public void testLongNotNotEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.notEquals(8l, 8l, "Failed")).isInstanceOf(
+        void testLongNotNotEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.notEquals(8l, 8l, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testLongNotEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.notEquals(2l, 3l, "Failed"));
+        void testLongNotEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.notEquals(2l, 3l, "Failed"));
         }
 
         @Test
-        public void testByteNotEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.notEquals((byte) 6, (byte) 6, "Failed")).isInstanceOf(
+        void testByteNotEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.notEquals((byte) 6, (byte) 6, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testByteNotEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.notEquals((byte) 5, (byte) 3, "Failed"));
+        void testByteNotEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.notEquals((byte) 5, (byte) 3, "Failed"));
         }
 
         @Test
-        public void testShortNotEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.notEquals((short) 6, (short) 6, "Failed")).isInstanceOf(
+        void testShortNotEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.notEquals((short) 6, (short) 6, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testShortNotEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.notEquals((short) 6, (short) 3, "Failed"));
+        void testShortNotEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.notEquals((short) 6, (short) 3, "Failed"));
         }
 
         @Test
-        public void testDoubleNotEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.notEquals(6.0, 6.0, "Failed")).isInstanceOf(
+        void testDoubleNotEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.notEquals(6.0, 6.0, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testDoubleNotEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.notEquals(5.0, 3.0, "Failed"));
+        void testDoubleNotEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.notEquals(5.0, 3.0, "Failed"));
         }
 
         @Test
-        public void testFloatNotEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.notEquals((float) 5.0, (float) 5.0, "Failed")).isInstanceOf(
+        void testFloatNotEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.notEquals((float) 5.0, (float) 5.0, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testFloatNotEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.notEquals((float) 3.0, (float) 4.0, "Failed"));
+        void testFloatNotEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.notEquals((float) 3.0, (float) 4.0, "Failed"));
         }
 
         @Test
-        public void testCharNotEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.notEquals('a', 'a', "Failed")).isInstanceOf(
+        void testCharNotEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.notEquals('a', 'a', "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testCharNotEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.notEquals('x', 'f', "Failed"));
+        void testCharNotEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.notEquals('x', 'f', "Failed"));
         }
 
         @Test
-        public void testBooleanNotEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.notEquals(true, true, "Failed")).isInstanceOf(
+        void testBooleanNotEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.notEquals(true, true, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testBooleanNotEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.notEquals(false, true, "Failed"));
+        void testBooleanNotEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.notEquals(false, true, "Failed"));
         }
     }
 
@@ -253,36 +283,36 @@ public class TestAssertionsEquals {
     class TestOptionalEquals {
 
         @Test
-        public void testOptionalObjectEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.optionalEquals(Optional.of("A"), "B", "Failed")).isInstanceOf(
+        void testOptionalObjectEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.optionalEquals(Optional.of("A"), "B", "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testOptionalObjectEqualsEmptyFail() {
-            assertThatThrownBy(() -> DomainAssertions.optionalEquals(Optional.empty(), "B", "Failed")).isInstanceOf(
+        void testOptionalObjectEqualsEmptyFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.optionalEquals(Optional.empty(), "B", "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testOptionalObjectEqualsFirstArgNullFail() {
-            assertThatThrownBy(() -> DomainAssertions.optionalEquals(null, "B", "Failed")).isInstanceOf(
+        void testOptionalObjectEqualsFirstArgNullFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.optionalEquals(null, "B", "Failed")).isInstanceOf(
                 IllegalArgumentException.class);
         }
 
         @Test
-        public void testOptionalObjectEqualsEmptyOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.optionalEquals(Optional.empty(), null, "Failed"));
+        void testOptionalObjectEqualsEmptyOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.optionalEquals(Optional.empty(), null, "Failed"));
         }
 
         @Test
-        public void testOptionalObjectEqualsOk() {
-            assertThatNoException().isThrownBy(() -> DomainAssertions.optionalEquals(Optional.of("A"), "A", "Failed"));
+        void testOptionalObjectEqualsOk() {
+            assertThatNoException().isThrownBy(() -> DomainAssertionsEquals.optionalEquals(Optional.of("A"), "A", "Failed"));
         }
 
         @Test
-        public void testOptionalObjectEqualsSecondArgNullOk() {
-            assertThatThrownBy(() -> DomainAssertions.optionalEquals(Optional.of("A"), null, "Failed")).isInstanceOf(
+        void testOptionalObjectEqualsSecondArgNullOk() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.optionalEquals(Optional.of("A"), null, "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
     }
@@ -291,31 +321,31 @@ public class TestAssertionsEquals {
     class TestOptionalNotEquals {
 
         @Test
-        public void testOptionalObjectNotEqualsFail() {
-            assertThatThrownBy(() -> DomainAssertions.optionalNotEquals(Optional.of("A"), "A", "Failed")).isInstanceOf(
+        void testOptionalObjectNotEqualsFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.optionalNotEquals(Optional.of("A"), "A", "Failed")).isInstanceOf(
                 DomainAssertionException.class);
         }
 
         @Test
-        public void testOptionalObjectNotEqualsEmptyFail() {
-            assertThatThrownBy(() -> DomainAssertions.optionalNotEquals(Optional.empty(), null, "Failed"));
+        void testOptionalObjectNotEqualsEmptyFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.optionalNotEquals(Optional.empty(), null, "Failed"));
         }
 
         @Test
-        public void testOptionalObjectNotEqualsOk() {
+        void testOptionalObjectNotEqualsOk() {
             assertThatNoException().isThrownBy(
-                () -> DomainAssertions.optionalNotEquals(Optional.of("A"), "Y", "Failed"));
+                () -> DomainAssertionsEquals.optionalNotEquals(Optional.of("A"), "Y", "Failed"));
         }
 
         @Test
-        public void testOptionalObjectNotEqualsEmptyOk() {
+        void testOptionalObjectNotEqualsEmptyOk() {
             assertThatNoException().isThrownBy(
-                () -> DomainAssertions.optionalNotEquals(Optional.empty(), "Y", "Failed"));
+                () -> DomainAssertionsEquals.optionalNotEquals(Optional.empty(), "Y", "Failed"));
         }
 
         @Test
-        public void testOptionalObjectNotEqualsNullFail() {
-            assertThatThrownBy(() -> DomainAssertions.optionalNotEquals(null, "A", "Failed")).isInstanceOf(
+        void testOptionalObjectNotEqualsNullFail() {
+            assertThatThrownBy(() -> DomainAssertionsEquals.optionalNotEquals(null, "A", "Failed")).isInstanceOf(
                 IllegalArgumentException.class);
         }
     }