+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.5] - 2025-02-07
+- Allow to pass in an own Exception instead of a String to DomainAssertions
+- Split DomainAssertions into separate classes for better maintainability
+
 ## [2.0.3] - 2025-01-29
 - Bugfix for proper handling of ValueObject duplicates in lists